requirements.txt
discord.py
mcstatus
websockets
python-dotenv